shrub_archi/requirements/install.txt
xmlschema
owlready2
rdflib
cst
defusedxml
msal
neo4j
networkx
pyopenssl
pillow
jinja2
dash
plotly